requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.2
mypy==1.0.0
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.0
requests-file==1.5.1
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
tweepy==4.12.1
types-beautifulsoup4==4.11.6.5
types-html5lib==1.1.11.11
types-requests==2.28.11.12
types-urllib3==1.26.25.5
typing_extensions==4.4.0
urllib3==1.26.14
